backend/requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.6
Flask-Cors==3.0.10
python-dotenv==0.21.0
